requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.29.1
schedule>=1.2.2
